mgk/requirements.txt
discord
jishaku
python-dotenv
datetime
typing
psutil
humanize
pytube
humanfriendly
pymysql
langdetect
translate
python-youtube
youtube_dl
